+++ lib/mmailer/mail_helper.rb
@@ -1,5 +1,6 @@
 module Mmailer
   class MailHelper
+    include Mmailer::Utilities
     attr_reader :template, :subject, :from
 
     def initialize(args)
@@ -22,12 +23,14 @@ module Mmailer
       mail.from = from
       mail.subject = subject
 
+      compiled_source=ERB.new(File.read(Dir.pwd + "/" + Mmailer.configuration.template + ".md.erb")).result(binding)
+
       text_part = Mail::Part.new
-      text_part.body=ERB.new(File.read(Dir.pwd + "/" + Mmailer.configuration.template + ".txt.erb")).result(binding)
+      text_part.body=compiled_source
 
       html_part = Mail::Part.new
       html_part.content_type='text/html; charset=UTF-8'
-      html_part.body=ERB.new(File.read(Dir.pwd + "/" + Mmailer.configuration.template + ".html.erb")).result(binding)
+      html_part.body=Kramdown::Document.new(compiled_source).to_html
 
       mail.text_part = text_part
       mail.html_part = html_part
@@ -36,11 +39,11 @@ module Mmailer
 
       case ENV['MMAILER_ENV']
         when "production"
-          mail.deliver!
+          try { mail.deliver! }
         when "development"
           puts mail.to_s
         else
-          mail.deliver!
+          mail.to_s
       end
     end
   end